Big time clean-up. A single user controller will suffice.
