+++ lib/mkoussaelixir_web/controllers/about_html/index.html.heex
@@ -1,8 +1,14 @@
 <section>
   <h1>ABOUT</h1>
   <p>Hello World! I'm Matthew Koussa and this is my site!</p>
-  <p>Here you'll find my forrays into digital synthesis via the <a href={~p"/loguesdk"}>Korg LogueSDK</a> or <a href={~p"/blorp"}>Blorp</a>, a synthesizer built using the p5 Javscript library.</p>
-  <p>Below is my <a href="https://mkoussa.bandcamp.com">Bandcamp</a>, where you can stream/download/buy my music!</p>
+  <p>
+    Here you'll find my forrays into digital synthesis via the
+    <a href={~p"/loguesdk"}>Korg LogueSDK</a>
+    or <a href={~p"/blorp"}>Blorp</a>, a synthesizer built using the p5 Javscript library.
+  </p>
+  <p>
+    Below is my <a href="https://mkoussa.bandcamp.com">Bandcamp</a>, where you can stream/download/buy my music!
+  </p>
   <p>This site is powered by:</p>
   <author>
     Elixir <%= Application.spec(:elixir, :vsn) %> | Phoenix <%= Application.spec(:phoenix, :vsn) %> | Hosted on
