--- lib/mkoussaelixir_web/controllers/user_confirmation_html.ex
@@ -1,5 +1,0 @@
-defmodule MkoussaelixirWeb.UserConfirmationHTML do
-  use MkoussaelixirWeb, :html
-
-  embed_templates "user_confirmation_html/*"
-end
