+++ lib/mkoussaelixir_web/controllers/user_html.ex
@@ -1,0 +1,5 @@
+defmodule MkoussaelixirWeb.UserHTML do
+  use MkoussaelixirWeb, :html
+
+  embed_templates "user_html/*"
+end
