--- lib/mkoussaelixir_web/controllers/user_registration_controller.ex
@@ -1,30 +1,0 @@
-defmodule MkoussaelixirWeb.UserRegistrationController do
-  use MkoussaelixirWeb, :controller
-
-  alias Mkoussaelixir.Accounts
-  alias Mkoussaelixir.Accounts.User
-  alias MkoussaelixirWeb.UserAuth
-
-  def new(conn, _params) do
-    changeset = Accounts.change_user_registration(%User{})
-    render(conn, :new, changeset: changeset)
-  end
-
-  def create(conn, %{"user" => user_params}) do
-    case Accounts.register_user(user_params) do
-      {:ok, user} ->
-        {:ok, _} =
-          Accounts.deliver_user_confirmation_instructions(
-            user,
-            &url(~p"/users/confirm/#{&1}")
-          )
-
-        conn
-        |> put_flash(:info, "User created successfully.")
-        |> UserAuth.log_in_user(user)
-
-      {:error, %Ecto.Changeset{} = changeset} ->
-        render(conn, :new, changeset: changeset)
-    end
-  end
-end
