--- lib/mkoussaelixir_web/controllers/user_registration_html.ex
@@ -1,5 +1,0 @@
-defmodule MkoussaelixirWeb.UserRegistrationHTML do
-  use MkoussaelixirWeb, :html
-
-  embed_templates "user_registration_html/*"
-end
