--- lib/mkoussaelixir_web/controllers/user_reset_password_html.ex
@@ -1,5 +1,0 @@
-defmodule MkoussaelixirWeb.UserResetPasswordHTML do
-  use MkoussaelixirWeb, :html
-
-  embed_templates "user_reset_password_html/*"
-end
