--- lib/mkoussaelixir_web/controllers/user_session_html.ex
@@ -1,5 +1,0 @@
-defmodule MkoussaelixirWeb.UserSessionHTML do
-  use MkoussaelixirWeb, :html
-
-  embed_templates "user_session_html/*"
-end
