--- lib/mkoussaelixir_web/controllers/user_settings_html.ex
@@ -1,5 +1,0 @@
-defmodule MkoussaelixirWeb.UserSettingsHTML do
-  use MkoussaelixirWeb, :html
-
-  embed_templates "user_settings_html/*"
-end
