+++ lib/mkoussaelixir_web/telemetry.ex
@@ -11,9 +11,9 @@ defmodule MkoussaelixirWeb.Telemetry do
     children = [
       # Telemetry poller will execute the given period measurements
       # every 10_000ms. Learn more here: https://hexdocs.pm/telemetry_metrics
-      {:telemetry_poller, measurements: periodic_measurements(), period: 10_000},
+      {:telemetry_poller, measurements: periodic_measurements(), period: 10_000}
       # Add reporters as children of your supervision tree.
-      {Telemetry.Metrics.ConsoleReporter, metrics: metrics()}
+      # {Telemetry.Metrics.ConsoleReporter, metrics: metrics()}
     ]
 
     Supervisor.init(children, strategy: :one_for_one)
@@ -107,7 +107,7 @@ defmodule MkoussaelixirWeb.Telemetry do
       # A module, function and arguments to be invoked periodically.
       # This function must call :telemetry.execute/3 and a metric must be added above.
       # {MkoussaelixirWeb, :count_users, []}
-      {Mkoussaelixir, :measure_orders, []}
+      # {Mkoussaelixir, :measure_orders, []}
       # {:process_info,
       #  event: [:mkoussaelixir, :my_server],
       #  name: Mkoussaelixir.Orders,
