+++ lib/mkoussaelixir_web/user_auth.ex
@@ -83,7 +83,7 @@ defmodule MkoussaelixirWeb.UserAuth do
     conn
     |> renew_session()
     |> delete_resp_cookie(@remember_me_cookie)
-    |> redirect(to: ~p"/")
+    |> redirect(to: ~p"/users/log_in")
   end
 
   @doc """
